requirements.txt
absl-py==2.0.0
astroid==3.0.2
attrs==23.2.0
black==24.4.2
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
descartes==1.1.0
diffusers==0.28.2
dill==0.3.7
efficientnet-pytorch==0.7.1
einops==0.7.0
exceptiongroup==1.2.2
filelock==3.13.1
fire==0.5.0
flake8==7.1.0
flatbuffers==24.3.25
fonttools==4.47.0
fsspec==2023.12.2
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
huggingface-hub==0.23.3
humanfriendly==10.0
idna==3.6
imageio==2.34.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.10.0
llvmlite==0.41.1
lyft-dataset-sdk==0.0.8
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.5.3
mccabe==0.7.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
nuscenes-devkit==1.1.11
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
oauthlib==3.2.2
onnxruntime==1.18.0
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.82
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
plotly==5.23.0
pluggy==1.5.0
pooch==1.8.2
protobuf==4.23.4
psutil==6.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyblaze==2.3.2
pycocotools==2.0.7
pycodestyle==2.12.0
pyflakes==3.2.0
pylint==3.0.3
PyMatting==1.1.12
pynvml==11.5.0
pyparsing==3.1.1
pyquaternion==0.9.9
pyro-api==0.1.2
pyro-ppl==1.8.6
pytest==8.3.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
rembg==2.0.57
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.1
Shapely==1.8.5.post1
six==1.16.0
sympy==1.12
tenacity==8.5.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2024.5.22
tomli==2.0.1
tomlkit==0.12.3
#torch==1.13.1
#torch-tb-profiler==0.4.3
#torchaudio==0.13.1+cu117
torchmetrics==1.2.1
#torchvision==0.14.1+cu117
tqdm==4.66.1
triton==2.3.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
xformers==0.0.26.post1
zipp==3.17.0

